--- src/features/contact/index.ts
@@ -1,1 +1,0 @@
-export { default as ChurchContact } from "./ui/ChurchContact";
