--- src/features/guide/index.ts
@@ -1,2 +1,0 @@
-export { default as TransportGuide } from "./ui/TransportGuide";
-export { default as ParkingGuide } from "./ui/ParkingGuide";
