+++ src/features/info/index.ts
@@ -1,1 +1,4 @@
 export { default as WorshipContact } from "./ui/WorshipContact";
+export { default as ChurchContact } from "./ui/ChurchContact";
+export { default as TransportGuide } from "./ui/TransportGuide";
+export { default as ParkingGuide } from "./ui/ParkingGuide";
